get and update admin profile APIs were moved to new class
